Delegate validClassFile() and toBinaryName() to SymbolLoaders.
The SymbolLoaders knows *how* to load class files. It is
therefore logical that it also has the responsibility to
tell *whether* a file can be loaded as class file or not.

Combined with the newClassLoader() hook, with this change
it is now easy to create a subclass of SymbolLoaders that
can load symbols from files other than .class files.