Make class loaders configurable by file extension.
The triplet validClassFile(), toBinaryName() and
newClassLoader() in SymbolLoaders does not hardcode
the .class file extension anymore.

Instead, SymbolLoaders keeps a list of mappings from
file extension to class loader factory. By default, it
contains the single mapping
".class" -> (bin => new ClassfileLoader(bin))
Additional mappings can be registered with
registerClassLoaderFactory().

With this change, it is not necessary anymore to create
subclasses of SymbolLoaders when we want to load
symbols from other file types.