WiP Remove the no-op jump from one case to the next in patmat codegen.
Previously, the patmat codegen generated code with the following
shape:

  val x = scrutinee
  case1() {
    if (x.isInstanceOf[A]) {
      if (deeperGuard) {
        matchEnd(...)
      } else {
        case2()
      }
    } else {
      case2()
    }
  }
  case2() {
    ...
  }
  ...
  caseN() {
    matchEnd(...)
  }
  matchEnd(...) { ... }

In all the possible ways that this code was generated, the jumps
to the next cases like `case2()` come in "tail position" before
automatically reaching the corresponding LabelDef.

That means that they can all be replaced by `()` (or an empty
`else` branch) without altering the behavior.

In this commit, we replace them with an empty `else` branch (i.e.,
an `EmptyTree`). This changes the type of the `If`s to `Unit`, and
all the way up to the `caseI()` LabelDefs themselves.

This change removes useless no-op `GOTO`s in the generated bytecode.